backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiosqlite==0.20.0
python-dotenv==1.0.1
anthropic==0.39.0
pandas==2.2.3
prophet==1.1.6
scikit-learn==1.5.2
numpy==1.26.4
